requirements.txt
discord.py
discord.py[voice]
psutil
asyncio
youtube_dl
aiohttp
mysql-connector-python